cql3: values: make raw_value fragmented
As a part of the effort of removing big, contiguous buffers from the codebase,
cql3::raw_value should be made fragmented. Unfortunately a straightforward
rewrite to a fragmented buffer type is not possible, because we want
cql3::raw_value to be compatible with cql3::raw_value_view, and we want that
view needs to be based on fragmented_temporary_buffer::view, so that it can be
used to view data coming directly from seastar without copying.

This patch makes cql3::raw_value fragmented by making cql3::raw_value_view
a variant of managed_bytes_view and fragmented_temporary_buffer::view.

Instead of treating those two view types as black boxes, we extract
their common parts (the first fragment and the size), which are most frequently
used, and store them directly in the raw_value_view. This way, we only need to
dispatch a call to the underlying view type when we are crossing a fragment boundary,
which is rare.