pgo: introduce PGO scripts
Profile-guided optimization consists of the following steps:
1. Build the program as usual, but with with special options (instrumentation
or just some supplementary info tables, depending on the exact flavor of PGO
in use).
2. Collect an execution profile from the special binary by running a
representative load on it.
3. Rebuild the program again, using the collected profile.

This commit introduces some scripts for running PGO training workloads on
Scylla, including one simple placeholder/example workload (cs-example).
We plan to replace this placeholder with a set of representative workloads
(ideally covering all hot code) in the future.

As input, the entry point script (pgo/run_all) takes a path to the special
binary, a path to a non-existing working/output directory, and the type of PGO
being used. As output, it produces the working/output directory containing a
profile file. The name of this file is a function of the PGO type.

In this commit, the only supported PGO type is LLVM IR-level instrumentation
(-fprofile-generate). At least one another type - BOLT instrumentation - will
be added in a future commit. Other possible relevang PGO types include
GCC instrumentation, LLVM source-level instrumentation, BOLT sampling and
LLVM CSSPGO, but we are unlikely to use them in foreseeable future.

The quality of PGO is correlated with the quality of the profile.
As a rule, the closer the profiling workloads are to the production workload,
the greater the gain from PGO.

Ideally all code that is hot in some production workload should be exercised
in the profile. Untouched code may be treated by the optimizer as cold
and e.g. get optimized for size (that's what the -pgso knob in LLVM does).
If this happens to actually hot code, it could in theory cause a performance
regression around that piece of code.

The ultimate profile can be merged (possibly with some weights) from multiple
profiles from multiple runs. Currently pgo/run_all delegates the profile
collection to workload subdirectories and merges the results without weights.

As of this commit, there is only one subdirectory - cs_example - included.
It loads Scylla with the base read/write workload of cassandra-stress.
This is just an example/placeholder; a good set of representative workloads
remains to be worked out.