db: config: add the bolt_instrumentation_dump_address option
Usually BOLT's instrumentation library dumps the result profile file to disk
on program exit. Scylla however skips the exit part and terminates itself via
a syscall, so the usual doesn't work. We have to call the dump function
ourselves.

This is awkward: BOLT injects the instrumentation library into an
already-linked program, so symbols from the library aren't visible in the
program. We only know the address of the dump function outside of the program
(because BOLT reports it while performing instrumentation) and we have to pass
it to the program somehow, so that it may call it.
There are many ways to do that. I chose argv as the most explicit one.

This patch adds a new option which asks Scylla to call a particular address just
before exit. This option is meant solely for internal use by our PGO build
scripts. They will instrument Scylla with BOLT, read the dump function
address from its stdout, and pass it to the instrumented Scylla when collecting
a profile from it.

Note that we are not hacking into BOLT's internals this way. Calling the dump
function manually using the address extracted from logs is supported, even
recommended.