test: boost: mulitshard_combining_reader_as_mutation_source_test: allow redundant range tombstone changes
test_range_tombstones_v2 is too strict for this reader -- it expects a
particular sequence of range tombstone changes, but multishard_combining_reader,
when tested with a small buffer, generates -- as expected -- additional
(redundant) range tombstone change pairs (end+start).

We would rather not modify test_range_tombstones_v2, because catching
redundancy is good in general, so to fix this problem, we wrap the
tested reader with a filter that removes these redundant range tombstone
changes, when they are expected.