scripts: add intel_pt_trace.py
Adds a script which sets up a cluster (via test/pylib),
runs something on it under Intel PT recording,
and converts the resulting trace to the FTF format, viewable
for example in Chromium's trace viewer (Perfetto).

It can be used to explore how some things in Scylla are implemented,
and how many instructions they take.
(However, without additional setup, it's only suitable for short workloads,
like a single query. Long operations, which, for example, rewrite system tables,
are likely to produce a trace that's too big for the tooling to handle.