sstables/mx/reader: handle inexact lookups in `advance_to_next_partition()`
After BTI, a generic `advance_to` will be inexact,
and might set the index reader to a data position earlier
than the actual target. This breaks the logic of
`advance_to_next_partition()`.

But if the searched partition is known to exist, BTI can
be used to find the exact result.
And in `advance_to_next_partition()` we know that the
partition exists. (Because we are currently in it).

So we can add a new API, `advance_after_existing(partition_key)`,
which, unlike `advance_to`, is guaranteed to give an exact result,
and we can use it in `advance_to_next_partition` instead of
of the `advance_to`.