compress: eliminate compressor_ptr from compression_parameters
`compression_parameters` is the piece of schema based on which
`compressor_ptr` objects are created.

Currently, it holds its own `compressor`, and new `compressor_ptr`
objects -- for writers, not readers -- are shared pointers to it.

Later patches make the `compressor` something more than a function
of a schema -- they add dictionaries, which aren't in the schema --
so this relationship isn't going to make sense anymore.

This patch does some refactoring to break the circular relationships
between compressors and compression parameters.
After the patch, compression_parameters don't hold a `compressor_ptr`.