mutation/tombstone: improve the printer
I recently had a test failure which reported that mutations are
not equal.
But in their printed form, they looked exactly the same.

It turns out that the inequality was caused by differing
deletion times in the partition tombstone.
But since both tombstones had a `missing_timestamp`, they were just
printed as `none`, not showing the different deletion time fields.

This patch improves the printer so that it only shows `none` if
the tombstone looks exactly what a live tombstone should look like.
If it's live, but the deletion time is something weird, the deletion
time will be printed.