sstables/trie: make bti_node_reader::page_ptr copy-constructible
In later commits, a trie cursor will be holding a `page_ptr`.
Sometimes we will want to copy a cursor, in particular
to do reset the upper bound of the index reader with
`_lower = _upper`. But currently `page_ptr` is non-copyable
-- it's a shared pointer, but with an explicit `share()` method
-- so a default operator= won't work for this.

Let's add a copy-assignable for `page_ptr` for this purpose.